Add dns types
